Initial Commit
HTML:
initial HTML body - includes a logo, the view and control sections:
input, priority and add button.
Below, the result is printed on the screen.

JS:
Created the add button function.
Created a JSON format element in order to save the data from the input.

CSS:
styling for the result padding and display only.