functions: append to local storage, counter, onReload(retrive the tasks after reloading the page)
append to local storage function - set the "myTodo" var as a key and the data in it as the value.

added counter - count the children of the "view-section" div: each child is a task.

added "onReload" func:  retrieve the tasks from the local storage and assaign it to myTodo var. append all the children with the task back to the main view-section