check and remove button, sort function improved
added 2 buttons: check and remove:
both buttons toggle class for the task's div. the check add line-through text decoration,
and the remove adds a display of none.

sort function: new function, much efficient and shorter witch rearrange the divs in the dom using for loops that check the priority value.