localstorage, buttons to a, counter improve and style
improve all functions to beeter communicate with the local storage when a task removal is occurred.

all buttons were replaced to a tag for style purposes.

counter function improved and the label is now changing if there are no tasks.

css styling:
titles to the task properties was added.
upgraded style for the input box and all the tasks divs.